package names lower cast, GPL license, google style
