+++ src/main/java/net/sf/mzmine/modules/visualization/kendrickmassplot/KendrickMassPlotModule.java
@@ -1,4 +1,23 @@
-package net.sf.mzmine.modules.visualization.kendrickMassPlot;
+/*
+ * Copyright 2006-2015 The MZmine 2 Development Team
+ * 
+ * This file is part of MZmine 2.
+ * 
+ * MZmine 2 is free software; you can redistribute it and/or modify it under the
+ * terms of the GNU General Public License as published by the Free Software
+ * Foundation; either version 2 of the License, or (at your option) any later
+ * version.
+ * 
+ * MZmine 2 is distributed in the hope that it will be useful, but WITHOUT ANY
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+ * A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License along with
+ * MZmine 2; if not, write to the Free Software Foundation, Inc., 51 Franklin St,
+ * Fifth Floor, Boston, MA 02110-1301 USA
+ */
+
+package net.sf.mzmine.modules.visualization.kendrickmassplot;
 
 import java.util.Collection;
 
@@ -11,6 +30,9 @@ import net.sf.mzmine.parameters.ParameterSet;
 import net.sf.mzmine.taskcontrol.Task;
 import net.sf.mzmine.util.ExitCode;
 
+/**
+ * Kendrick mass plot module
+ */
 public class KendrickMassPlotModule implements MZmineProcessingModule{
 
     private static final String MODULE_NAME = "Kendrick mass plot";
@@ -30,14 +52,14 @@ public class KendrickMassPlotModule implements MZmineProcessingModule{
     @Nonnull
     public ExitCode runModule(@Nonnull MZmineProject project,
             @Nonnull ParameterSet parameters, @Nonnull Collection<Task> tasks) {
-       
-    	  Task newTask = new KendrickMassPlotTask(parameters);
-          tasks.add(newTask);
-      
-          return ExitCode.OK;  
+
+        Task newTask = new KendrickMassPlotTask(parameters);
+        tasks.add(newTask);
+
+        return ExitCode.OK;  
     }
 
-   
+
 
     @Override
     public @Nonnull MZmineModuleCategory getModuleCategory() {
